+++ src/imio/helpers/xhtml.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_base
+from html import escape
 from os import path
 from plone import api
 from plone.app.imaging.scale import ImageScale
@@ -625,7 +626,7 @@ def separate_images(xhtmlContent, pretty_print=False):
                                        method='html') for x in tree.iterchildren()])
 
 
-def object_link(obj, view='view', attribute='Title', content='', target=''):
+def object_link(obj, view='view', attribute='Title', content='', target='', escaped=True):
     """ Returns an html link for the given object """
     href = view and "%s/%s" % (obj.absolute_url(), view) or obj.absolute_url()
     if not content:
@@ -636,4 +637,6 @@ def object_link(obj, view='view', attribute='Title', content='', target=''):
             content = content()
     if target:
         target = ' target="{}"'.format(target)
+    if escaped:
+        content = escape(content)
     return u'<a href="%s"%s>%s</a>' % (href, target, safe_unicode(content))